+++ src/components/Pulse/stories/PulsePlayground.template.tsx
@@ -1,0 +1,10 @@
+import { Pulse } from '@admiral-ds/react-ui';
+import type { PulseProps } from '@admiral-ds/react-ui';
+
+export const PulsePlaygroundTemplate = (props: PulseProps) => {
+  return (
+    <>
+      <Pulse {...props} />
+    </>
+  );
+};